ArchInterrupt：更改实现方式
